+++ Lib/Watcher/OsWatcher.cs
@@ -4,6 +4,17 @@ using Lib.Utils;
 
 namespace Lib.Watcher
 {
+    public class DummyWatcher : IDirectoryWatcher
+    {
+        public void Dispose()
+        {
+        }
+
+        public string WatchedDirectory { get; set; }
+        public Action<string> OnFileChange { get; set; }
+        public Action OnError { get; set; }
+    }
+
     public class OsWatcher : IDirectoryWatcher
     {
         public string WatchedDirectory